pyproject.toml
[tool.poetry]
name = "robotframework-web-example"
version = "1.0.0"
description = "Proyecto de scripts automatizados con Robot Framework y SeleniumLibrary"
authors = ["Angel Gerardo Molina Valdez <angelgerardomolinavaldez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
robotframework = "^7.0"
robotframework-seleniumlibrary = "^6.2.0"
pyyaml = "^6.0.1"
robotframework-robocop = "^5.0.3"
jinja2 = "^3.1.3"
robotframework-datadriver = "^1.11.0"
pillow = "^10.2.0"
reportlab = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
